Fix fragment index
